NDMP: fix restore problems with jobs with multiple BSRs
The handling of multiple BSRs (Bootstrap Records) was not
correct for restores.

When an NDMP job is stored over multiple media, for
each medium is a BSR created and the old code also created a
individual restore job for each medium which is not correct.

The new code creates a restore job for each SessionID - FileIndex
tuple which identifies an individual NDMP stream and as such needs
and own restore job.