Windows installer: workaround to keep config files
The Bareos Windows installer versions 16.2.4 and 16.2.5 remove legacy
configuration files (eg. bareos-fd.conf), even if "keep old
configuration" is selected. This bug has been fixed, however, when
upgrading from one of this versions, the uninstaller of this versions
still removes the legacy configuration files.
With this commit, the upgrade process backups of the existing configuration files before
calling the uninstaller. After the uninstaller, the configuration files
are copied back to their original location.