list files command: show deleted files
until now, the "list files jobid=" command did
simply show the files that were backed up in a certain
jobid.

As in accurate mode, the files are send by the fd as "backed up" but marked as
deleted (FT_DELETED). This information is then stored in the database by setting the
fileindex of that file as 0.

With this patch, the deleted files are visible as such in the output of
"list files jobid=" as the following output shows.

First a backup was done of a small test fileset:

*list files jobid=1
 /testdata/testdir/testfile
 /testdata/testdir/
 /testdata/testdir2/
 /testdata/testdir2/testfile2
 /testdata/

then, testdir2 and its content was deleted before running the
second incremental backup:

*list files jobid=2
 /testdata/testdir2/ (deleted)
 /testdata/testdir2/testfile2 (deleted)
 /testdata/

The output of the command shows that both the directory and the file
have been detected as missing and marked as deleted by appending a

       "(deleted)"

after the file/directory name