remove deprecated readdir_r() calls
since glibc version 2.24, readdir_r() is deprecated.

From the manpage of readdir_r():

It  is  expected that a future version of POSIX.1 will make
readdir_r() obsolete, and require that readdir() be thread-safe
when concurrently employed on different directory streams.

By defining "USE_READDIR_R" the old code is still available
in case it is needed on certain platforms.