output_formatter: use Dmsg instead of Emsg
In some environments, these error messages are triggered often, without
being actual errors (client send command to Director and closes the
connection. In this case, the Director can't send the result back to the
client and issues this error).
With this patch, a debug message is issued instead of an error message.