gfapi: Don't reinitialize the connection to gluster
The plugin event bEventRestoreCommand is fired for each file we are
supposed to restore, currently we don't check if we already have setup
the connection to gluster using the function connect_to_gluster() and as
such we leak a set of socket connections for each file restored. This is
bad so lets properly check if things are already setup correctly.

For a backup we do the same e.g. check if p_ctx->glfs is not already
initialized.