stored: Use SEEK_CUR not SEEK_END to check positioning.
When we want to determine if the position is set right in
DCR::is_eod_valid() we need to perform a lseek(fd, 0, SEEK_CUR)
and not a lseek(fd, 0, SEEK_END) as that is already done in
DEVICE::eod(). We only want to determine if the lseek was done
right in DCR::is_eod_valid() and by doing a SEEK_END we do the
same seek twice which leads to the same result but is not correct.