lib: Add edit_pthread() function to print pthread_t ids.
There is no real clean way to print the pthread_t type returned by
various pthreads functions. Until now things were printed as pointer
addresses. The edit_pthread() method is analog to other edit functions
and tries to print things in an uniform way and should be used to print
the content of the pthread_t opaque structure. The benefit of this is
that the representation is in one central place and we can implement
specific representations for different platforms.

As part of the implementation of this new function the edit.c source
file was also refactored and reindented to be better readable. Further
more now all modifiers are stored as static const char pointers which
used to be done for some functions but also not in others.