Fix to bugreport #840
To allow end users to choose the MPM way of running php7 the package
should only require httpd (work also for RHEL related) and one php
engine.

It can recommend mod_php but this will not last long in the future.

Note:

SUSE's RPM supports the Recommends tag, but it was not available
upstream until very recently. As a result, packages with Recommends will
fail to build for OBS targets Fedora_18, RHEL_6 and CentOS_6, for
example. A workaround is to exclude the Recommends tag from non-SUSE
targets. Alternatively the Recommends tag could become a Requires tag so
that a recommended package (foo) gets installed in any case on systems
that do not support RPM Recommends.

Fixes #840: bareos-webui openSUSE RPM has hard require to apache2_mod-php (5 or 7)