NDMP: add support for 2-way NDMP with DAR and DDAR
The current implementation of NDMP in Bareos has some limitations
because of the way it was implemented.

- The backup is always sent over the network from the Data Agent to the
  Tape Agent (3-way NDMP)
- DAR(Direct Access Recovery) and DDAR (Directory DAR)  are not supported,
  single file restores are VERY slow as the whole backup stream needs to
  be scanned during restore.

This commit adds support for 2-way NDMP backups inside of the Director.
During 2-way NDMP Backups, the Director is the only bareos daemon involved,
the Storage Daemon does not do anything in this setup.

To be able to do 2-way NDMP backups and to support DAR and DDAR, the
director
- controls the NDMP changer via NDMP Protocol.
- does the bookkeeping for the NDMP media.
- stores the full NDMP file information being sent during backups (FhInfo and FhNode)
  in the catalog database.
- sends back the full NDMP file information during restore to enable DAR
  and DDAR.

The advantage of the 2-way NDMP backups are:
- Much higher backup speed as the data goes directly from the disks to
  the tape without going over the network.
- Much higher restore speed for single files/directories with DAR and
  DDAR, as only the required blocks are read during restore instead of
  restoring the whole backup stream and only restoring the required
  data.

The following NDMP Protocol keywords are now available:

 * NDMP       :  which is an alias for NDMP_BAREOS for backward compatibility
 * NDMP_BAREOS:  which is NDMP backups with paired storage and NDMP Tape
                 Agent inside of the  Bareos Storage Daemon
 * NDMP_NATIVE: this is doing NDMP backups with only the DMA inside of
                the Bareos Director without any connection to the SD

These keywords reflect the two ways of NDMP that are now supported in
Bareos, which is either the special way of NDMP backups implemented in
Bareos before, or the "NATIVE" way to do NDMP backups.

To be able to store the FileHandle information for DAR and DDAR, the
File table how has two additional columns for FhInfo and FhNode.
The JobMedia table how has the additional column JobBytes which is used
to store NDMP media Information.

This column is planned to hold the JobBytes that are stored in the
corresponding JobMedia record.

The ndmp_dma_backup and ndmp_dma_restore files were split up for the two
different ways of NDMP backups that we have now.