Reset JobStatus to previous JobStatus in status SD and FD loops
We need to reset the JobStatus value to its previous after each storage or client status call,
otherwise if we receive an JS_Error for one storage or client, all the following status calls
will fail as well.

This could happen for example if a client or storage is offline but more status calls to other
clients and storages will follow in the loop.