Clients: Tiny Client Details View Improvement
Be more precise about what is displayed in the client details
'last backups' table. Therefore the heading is renamed to
'last successful backups' and an additional row 'Status',
which displays the JobStatus, is added to the table.

Fixes #812: Add job status in the "Last Backups" section of the Client view