stored: Fix chunked_device writing.
We missed a corner case where the write exactly aligns with the
chunksize. Also added a guarding MIN() in the overflow writing part so
we never write more then count bytes and crash in memcpy() on to many
bytes to copy for the buffer being passed in.

The code now has more defensive coding and some things might never be
triggered but better make sure things always work. You can now also have
a blocksize that is bigger then your chunksize (wrong config but
possible to configure) and then still things should work.