Build: switched to cmake as build system
Bareos now uses cmake as build system instead of
the old autotools system.

All c++ files were renamed to *.cc so that it is clear
which files need to be compiled wit the c compiler and
which ones need to be compiled with the c++ compiler.

Switching to cmake saved us 78277 lines of code while only
adding 4983 lines of code.

The build system is much clearer with cmake and the builds
also build faster than before.

Thank you to the people at Kitware for creating cmake!