bareos-storage-droplet: renamed bareos-storage-object
The name bareos-storage-object has been very generic.
This backend utilize the the droplet library
and even so it has been implemented to store data on S3 Object Storages,
in can in principle also be used for other storages support by droplet.

For this reason and to indicate the close relation to droplet,
the source files and the packages have been renamed.