Plugin option vcthumbprint optional
As of VDDK 6.5, passing the SSL certificate thumbprint is mandatory.
For ease of use, the vcthumbprint plugin option is now optional, if it
is not given as a plugin option, it will be retrieved from the vCenter
server. This is less secure, but will just work in most cases.

Also added the optional plugin option transport was added, which can be
used to force bareos_vadp_dumper to use the specified transport method.

The plugin and vmware_cbt_tool.py now disables SSL certificate
verification in python versions >= 2.7.5, this was >= 2.7.9 before, but
CentOS/RHEL 7.4 changed the python default to verify SSL certificates
while the python version stayed at 2.7.5