stored: Fix device not being closed problem.
When we cannot read any volume header from a volume for whatever reason
the current code will not re-open the device when retrying to mount the
next available volume in the DCR::mount_next_write_volume() method. This
fixes the code to explictly do a close on the device when we are
retrying to mount the next volume and its not a tape device and the
device is still open just before we call the autoload_device() function.