stored: Only pass the data pointer to the update function.
We do not have to expose the internal structure used on the ordered
circular list when calling the update function. We more or less cloned
the compare function which is used by the dlist::binary_insert()
method which operates on the actual list item and not on the data
contained in that list item.