stored: Fix reading from a still inflight chunk.
When we use io-threads we could end up in the situation where the data is
still queued to be uploaded to the backing store while a new backup starts
and will open the volume again and try to read the volume label. This used
to fail as we always only looked at the backing store for the source of
a volume chunk which was not available yet. We now clone the data from
the ordered circular buffer back to the current in memory chunk buffer
which gets newly allocated on re-open of the device and we also track
any inflight data to the backing store e.g. chunks currently being
uploaded, for such chunks we busy wait until they are uploaded or when
uploading fails are placed back on the ordered circular list with pending
chunk io requests.