stored: Limit retries to flush to backing store.
Make it possible to configure an upper limit on the number of tries that
the chunked volumes abstraction uses to upload the data to the backing
store. When this number of tries is exceeded the chunk is dropped and
the device is put into read-only mode not allowing any more writes or
(re)opens with the write flag set.