stored: Fix getting size of a chunked volume.
We could end up in trying to get the current volume size from the
backing store while the last chunk of the volume was still inflight
to the backing store. Previously we added support for getting the
size from the ordered circular list but that is only part of the time
it takes for a volume to get flushed to the backing store. When the
last volume chunk is still inflight we get the wrong size when we
list the volume chunks on the backing store as the chunk might be
still growing in size while being uploaded.