purge: fixes ActionOnPurge
ActionOnPurge have been non functional since bareos-17.2.

This patch make it work again.
Additional, it fixes a newly introduced bug,
that a "purge volume=VOLNAME ... action" command
truncates all volumes of a specific storage and pool and not only on the volume given as parameter.
As the volumes have been already purged, this have not been a big
issue.

Note: it is advices to use the "truncate" command for truncating and not
the "purge ... action" command.