redirect stderr from xtrabackup to logfile in WorkingDirectory
this doesn't seem to work:
(self.subprocess_stdOut, self.subprocess_stdError) = self.stream.communicate()

instead introduce a new parameter, log, which defaults to "plugin-percona.log".
a relative path will be relative to WorkingDirectory, an absolute path is allowed.
it can also have the value "false" to send stderr to /dev/null

the log file is opened for writing (ie. it will truncate an existing file)
so it is not likely to fill the disk.  in my opinion the log file is only useful
for debugging, so historic log data is not of interest.