HTML: fixes some more problems on copy-and-paste
In the HTML version of this document, some characters in the
listings have been transformed to look better (ligatures like 'fi' or 'fl'
into a single character 'fi').
However, these characters did cause problems when using
copy-and-paste.

As the documented ways did not work to disable this behavior, we
have to manually create a list of characters, that should not be
transformed.

We added a set of ligatures that should not be transformed.
We hope, we have now covered all character groups that could cause
problems.