changes to configuration to get tls-psk work
As tls-psk checks accurately name and password during,
it is no longer allowed to have different names between
the clients' own name and what name is configured for that
client in the director.

Before, the director called the local client as "bareos-fd"
while the client itself was automatically called after the
hostname of the local machine as ${hostname}-fd.