bareos-storage-droplet: improve error handling
Previos version did not always return an error, if data could not be written.
Especially the load_chunk ignored EIO errors, properly because of a typo.

As the droplet_device in iothread mode relies on asynchronious write-backs,
the new device method flush() has been introduced.
If a droplet_device is configured to use iothreads and unlimited retries,
this will do busy waiting until all data is written to the droplet backend.
In case of a connection problems to the droplet_device, this will be forever.

Note that a bconsole "status storage=..." command will inform about "Pending IO flush requests".

Fixes #892: bareos-storage-droplet: if configured with unreachable S3 system, backup will terminate with OK