Fix gfapi-fd: handle non-fatal Gluster problems properly
Previously when the gfapi-fd plugin encountered a stale file handle
problem in Gluster, a job error message was generated and the processing
of the glusterfind list stopped but the job terminated with status
Backup OK -- with warnings.

Also when using the glusterfind and the gffilelist option, files in the
list but missing on disk have been silently ignored. Now, a warning
job message is emitted in this case but not causing the termination
status with warning.

With this change, stale file handle problems are considered non-fatal,
for each such file an error is reported in the job messages and
and the termination status will be Backup OK -- with warnings.
When files that were in the glusterfind list are not found when they
going to be backed up, a warning is reported in the job message but
the termination status will be just Backup OK (not with warning).
Any other error is considered fatal and the job will terminate with
the Backup Error status immediately.

When gluster has a stale file handle problem, the plugin crashed
with a segmentation fault when not using the glusterfind option.
This is fixed now by skipping useless calls of checkChanges() in
that case.