sqlqueries: fix core/src/cats/dml/create_queryfiles.sh
The name of the BareosDB class was changed but the query create
script was not adapted to that fact.

This is now fixed and the queries were created anew.

Also changed GetQuotaJobbytes and GetQuotaJobbytesNofailed back
to the old name because that is the same name as the query file has.