stored: do not treat read error as EoT by default
Fixes #1034: Read error on tape may be misinterpreted as end-of-tape.

Previous storagedaemons treated a read-error immediately following
an end-of-file mark on a tape as end-of-tape instead of an error.

This patch makes storagedaemons raise an error in this case, but
allows to switch back to the previous behaviour on a per-device
basis using the new configuration directive "Eof On Error Is Eot".

Documented new configuration directive.