vmware: fixed errors when using non-ascii characters
Fixed UnicodeEncodeError in the Bareos VMWare Plugin when the VM name
and/or VMDK file name or folder name contained non-ascii characters.
It is also possible now to configure backups of VMs contained in vApps
by handling the vApp name like a folder name.

The script vmware_cbt_tool.py was also adapted to handle non-ascii
characters and vApps properly. The --listall flag was added to output a
list of all VMs in the given datacenter including the instanceUUID,
folder name and VM name. Also it's possible now to specify the VM by
using --vm-uuid <instanceUUID> instead of folder and VM Name.

Additionally small changes to comply with flake8 --max-line-length=100.