regress: first commit to modernize regress tests
We will move the exising "regression" tests test by test
into the "systemtests" directory finally removing the
regress folder completely.

The systemtests tests are part of the cmake project
and so can be executed with "make test" directly.

The tests run on the compiled binaries inside of the
CMAKE_BINARY_DIR without needing a "make install" at all.

The idea is to have an individual folder inside of systemtests
for every test that has the tests name.

The first example for how that works is the
systemtests/tests/backup-bareos-test folder.

- Each test folder needs to be indipendent from the other, so that
  ctest can execute each of the tests in parallel.

- The full configuration for each test is stored in etc/bareos
  as .conf.in template files that are edited by cmake during the cmake
  run.

- each test has a "testrunner" script file that executes the test
  itself.

- The "testrunner" sources an individual environment shell script that
  contains the variables that are used for this test.