docs: (Sphinx) store converted images
In the LaTex documentation, the images have been stored in the
docs/manuals/images/ directory.
During build of the documentation,
the images have been converted to the appropiate formats
and temporarly stored in the
docs/manuals/images/conv/ directory.

The current build process to not allow the conversion step
and therefore the relevant images are now also stored in
docs/manuals/en/new_main_reference/source/_static/images/
directory.