dird: redesign GetNdmpEnvironmentString() API
Fixes #1056: NDMP restore on 16.2.5 and above does
             not fill NDMP environment correctly

Previously one overload of the function
GetNdmpEnvironmentString() wanted a JobDbRecord*
and expected jr->VolSessionId and jr->VolSessionTime
to contain the values for the volume from which the
restore happens. These had to be filled manually
before calling GetNdmlEnvironmentString() which had
not been done since 16.2.5 resulting in #1056.

This patch now redesigns the API for all overloads
of GetNdmpEnviromentString() to make it harder
to misuse.

We also add a new struct VolumeSessionInfo to wrap
a pair of VolumeSessionId and VolumeSessionTime.
These two numbers are only meaningful together, so
they now have their own container.