cats: handle empty/non-empty environment correctly
The previous version did not detect whether the
NDMP environment contained values or not.
This new version correctly detects the amount of
environment values returned from the database
and returns success/failure as expected by the
calling code (i.e. return false when there is no
environment in the database).