stored: do not write duplicate SOS during MAC run
Fixes #1030: bscan does not work for migration and copy jobs

Previously when running a migrate, copy or virtual full backup the
resulting volume contained not one but two SOS records. The newly
generated one and the original one.

This patch now makes sure that only the newly generated SOS record is
written to the destination volume.
This patch also makes sure that the fields job_name, client_name,
fileset_name and fileset_md5 are written in the new SOS.