webui: Fix overflowing top navigation bar content
Since Bootstrap doesn't know how much space the content in a navbar needs,
it can cause issues with content wrapping into a second row.

To resolve this, we can:

1. Reduce the amount or width of navbar items.
2. Hide certain navbar items at certain screen sizes using responsive utility classes.
3. Change the point at which the navbar switches between collapsed and horizontal mode.

As we don't want to fiddle with SAAS to customize the Bootstrap @grid-float-breakpoint
variable, we simply add our own media query.

By default the collapse occurs before 768px. This patch overrides the default behaviour
with a custom media query, so the collapse happens earlier at 1330px to avoid the top
navigation bar overflowing issue.