webui: Display additional information before restore form submit
Replaces the JS confirm message with a bootstrap modal. The modal
displays the choosen restore parameters for user verification
before submit and the restore job being scheduled.

The commit also adds another button, named "Job report", to the
director message modal, which is displayed after form submit and
the restore job being scheduled.

The JS alert message, that is shown when no files or directories
have been selected, is replaced by a bootstrap modal as well.