cats: Fix .bvfs_lsdirs when using limit and offset
Fixes #975: .bvfs_lsdirs limit offset command parameters do not work properly

The .bvfs_lsdirs command did not return the correct amount of entries
when it was used with multiple jobids together with limit and offset.
The problem was that the SQL query returned multiple rows with the same
path and Bvfs::_handle_path() removed the duplicates. Also, no matter
what offset was given, the result did always start with the . and ..
entries.

This is fixed now when using the PostgreSQL DB for Catalog by a modified
SQL query that integrates the formerly separate SQL query for the . and
.. entries with the query for the normal directory entries and the new
query no longer contains duplicates.

This version will not work with a MySQL Catalog DB because MySQL/MariaDB
lacks the DISTINCT ON feature.

Original Author: Stephan Duehr <stephan.duehr@bareos.com>