webui: Notify user about Bvfs cache update
Display an bvfs update cache notification in the restore filetree
container while loading the tree and in case of an error after tree
loading failed.

The filetree refresh timeout is now configurable by the
filetree_refresh_timeout parameter in the restore section of the
configuration.ini file. The default of the parameter is
set to 120 seconds.

[restore]
; Restore filetree refresh timeout after n milliseconds
; Default: 120000 milliseconds
filetree_refresh_timeout=120000