dird: fix Pool_ACL checking in the "llist media all" command
Checking the Pool_ACL in the "list media all" is not possible,
therefore the prevent the command execution when a Pool ACL is set.

Checking the Pool_ACL in the "llist media all" command
no works as expected.
Changes: the result now also contains the columns "Pool", "ScratchPool" and "RecyclePool".