Add prefix-map compile options
GCC starting with version 8 supports a flag called -fmacro-prefix-map
that allows you to remap source paths. This will affect the __FILE__
macro.
This patch remaps the core directory to . so we leak fewer build system
paths into the final binaries.

This patch also adds -fdebug-prefix-map parameter to the compiler calls.
This is usually something you want which is why it is enabled by
default (and maps the same was as -fmacro-prefix-map).
To disable the behaviour the new cmake option DEBUG_PREFIX_MAP has been
added. This can be set to NO to disable remapping of debug source paths.

The patch also disables DEBUG_PREFIX_MAP for RPM and DEB builds, because
these build systems rewrite the paths themselves and thus need them
unchanged.