Allow forcing the version with -DVERSION_STRING=
previously the scripts get_version.cmake and write_version_files.cmake
did not honor a VERSION_STRING that was set on the commandline. Now
you can use cmake -DVERSION_STRING -P <script>.cmake to force the
version to be used.
While this is not very useful in case of get_version.cmake it will be
used to write the correct release version numbers into the release
tarballs.