Fix restore of thin-provisioned disks to VM
The restore via imageio requires the exact number of bytes
to be transferred, this is now implemented by saving the
number of bytes backed in restoreobjects.

Also fixed a typo so that getting the correct actual_size from
OVF data now works.

Finally added transfer rate statistics for each disk image
in the job log.