fd-plugins: Check return value of endRestoreFile
When omitting return bRCs['bRC_OK'] in the end_restore_file method
of a Python plugin, it resulted in seemingly unrelated Python
errors like "TypeError: an integer is required" in subsequent calls
of Python funtions.

Now the return value is checked and if it is invalid, an appropriate
error message will be generated.