core: include BareosLocalBuildDefinitions later
Previously BareosLocalBuildDefinitions has been included after
determining the version. This made it impossible to override a lot of
things. For example, you could not reconfigure a target with the local
definitions.
This patch moves the inclusion after the inclusion of subdirectories.
This allows you to override virtually anything that is configured inside
the core/ directory with your local definitions.

The patch also removes BareosLocalBuildDefinitions.cmake as it is not
required, but an optional local addition.