packaging: disable debtransform, remove *.changes
Nowadays Jenkins will select the release number we want to build, so
we disable the debtransform mechanism that overwrites the version.

Jenkins also generates a changelog that refers to the release notes.
Thus we should remove the unmaintained .changes files.
As Travis used the changes file to generate a debian changelog, we
now generate a dummy changelog.