packaging: remove bareos-release.key
This file is only meaningful if it matches the key the binary-repository
is signed with. Thus the file should be injected by the build process,
that will know how the packages are going to be signed.
The patch also adds information how to obtain the file
bareos-release.key manually if needed.

Co-Authored-By: Jörg Steffens <joergsteffens@users.noreply.github.com>