dird: fix RunScript parsing
This commit fixes a bug introduced during the recent months.
Without this patch, all RunScripts are configured to "Runs On Client = Yes".
Because of this, Admin and Console RunScripts did not work and are silently ignored.

With this change, RunScripts can by set to run locally or on the Client (Runs On Client = Yes/No).