cmake: Use date command on cmake <3.10 so %B works
Previously cmake fell back on using the date command in favour of
TIMESTAMP when the cmake version was 3.8 or older. This lead to a case
where a literal %B was added in the Bareos date when building with
cmake 3.8 to 3.10.
This patch makes cmake fall back to use the date command on versions up
to 3.10.