cats: Fix upgrading MySQL database from version 2171 to 2192
"CREATE INDEX IF NOT EXISTS ..." does not work on MySQL.

Instead "ALTER TABLE Job ADD INDEX (JobTDate);" is compatible to
MariaDB and MySQL. This also works if an index is already in place.

Unfortunately we could end up having multiple indexes on
JobTDate this way, so manually cleaning up could be required.

Perform also a key distribution analysis on the PathVisibility
table after index drop.

Fixes #1176: Upgrade 18.2.5 to 19.2.5 failed ...