lib: make IsAclEntryValid() more robust
Fixes #1175: bareos crashes with invalid character in ACL

Previously IsAclEntryValid() took a char* as a parameter that was
expected to be a POOLMEM (and would have been resized when an error
occured).
This char* has been replaced by a std::vector<char>& and the new
overload for Mmsg() is now used to format a message.