webui: Fix sorting of job selection in restore module
A recent bconsole command change in the getClientBackups method
of the client model lead to a wrong population of the backup job
selection to choose from in the restore module.

To fix this issue the population of the backup job selection list is
still done by the getClientBackups method but with the proper bconsole
command as it has been before.

The client details action now calls the getClientJobs method of the
client model to get a job list with all status instead a list of
successful backup jobs only.

Fixes #1174: reversed sort order on the restore client selection