tests: test crc32 with a real label block
Bug #1180: CRC checksum algorithm changed between 18.2.7 and 19.2.5

Previously the crc32 tests did only rudimentary changes, but did not
check with a real bareos block. This patch now adds a dumped label block
from a test-installation and calculates the checksum for that.
This patch also changes the pattern for another test, so it triggers on
an endianess problem too.