lib: allow more chars in IsAclEntryValid()
Bug #1175: bareos crashes with invalid character in ACL

The newly introduced IsAclEntryValid() allowed only "!*.:_-'/" as chars
in an ACL. As regular expressions are possible in ACLs this list did not
allow some previously valid ACLs to be configured.
This patch extends the list to allow "!()[]|+?*.:_-'/" which will allow
most syntaxes.
We do not add "{}," intentionally, because this would re-allow the wrong
ACL syntaxes that the original change tried to catch (i.e. adding a
string with comma-separated resource names instead of a list of resource
names).