stored: initialize jcr->sd_tls_policy
Fixes #1192: Authorization key rejected by Storage daemon since
upgrading director and storage daemons

Previously the sd_tls_policy member of the JCR in the sd was not
initialized properly. By using the uninitialized value, the SD may
close a valid non-TLS connection which eventually leads to a failing
job.
This patch initializes sd_tls_policy to kBnetTlsNone (0) which lets the
SD accept every connection on a job that was not configured otherwise by
the director.