droplet: only allow configuration maximum concurrent jobs of 0 or 1
There is a known bug when using Storage Backend Droplet with multiple simultaneous jobs (interleaving).
This workaround only allows settings of 0 or 1.
If the default is used (0), it will be set automatically to 1.
If any other value is specified, the bareos-sd will not start at all.

Check droplet devices immediately after configparser-run