core: add infrastructure for configuration warnings
This patch adds a list of warnings to the configuration parser along
with some utility functions.
The director, storagedaemon and filedaemon in config_test mode will
show this list of warnings.
When using the status commands, the daemons will report if they have at
least one record in that list.
The director now has a new command "status configuration" that will
currently show only that list of warnings.