core: use stdint.h where possible
Previously bc_types.h declared a lot of types itself. As of C99 the
compiler provides these types if they are available on the platform, so
we don't need any tricks for this anymore.
This patch removes most of the fixed-width integer aliasing in favor of
the standard-provided fixed-width types.