filed: remove support for bVarDistName
The plugin interface allowed to query the platform, distribution and
version of the filedaemon using bVarDistName. This has not been used by
any plugin and the provided value was of questionable quality.
This patch removes support for bVarDistName. Querying it will now raise
an error. The related dist_name from message.cc that was used solely for
this purpose has also been removed.