core: provide GetOsInfo() in kBareosVersionStrings
This patch adds a new function GetOsInfo() to kBareosVersionStrings. In
the future this function will be used as a replacement of the DISTVER
definition.
Eventually the GetOsInfo() should retrieve the information at runtime
instead of having it compiled in.

This patch also replaces DISTVER with a call to
kBareosVersionStrings.GetOsInfo() where possible. For windows this means
that we now always report the information obtained at runtime when the
program started.