Update login.phtml
Hides a the login form and shows a "loading" message while it does the Ajax call to get the versions. This avoids the problem that happens when you load the login form and click login too quickly (it logs in incorrectly) and also the visual effect of the dirctor and locale empty selects.