core: sort jobids before purging
previously the list of jobs was not sorted when the prune volume code
ran. This patch sorts the list so the log message and SQL query are
more predictable.
This also changes some prototypes to const where possible (and required
by std::string).
Lastly we add a "wait" in the volume-pruning-test to make sure the
previous jobs have finished before we start the jobs that prunes.