run command: use specified since time
The option "since" already existed in the run command, but was not
used anywhere.

Now it is possible to run a job in the following way:

run job=backup-bareos-fd since="2020-04-01 11:11:11"

This will run the given job and use the given since time.
No further calculations like checking for previous backups or considering
the backup level.

The jcr member stime was renamed to starttime_string to make it clear
that a string contaning a timestamp in text form is stored here, in
opposite to a string containing a unix timestamp.