systemtests: always reset dbdriver before setup
Previously once setup by the setup-script the catalog for the systemtest
was set to a specific database type. This made it impossible to test
different database backends without reconfiguring the tests in between.
This patch now always resets the dbdriver setting in the catalog
configuration so setup always changes dbdriver to the correct value.